Add BaseStore to avoid some of the boilerplate
