+++ app/stores/BaseStore.js
@@ -1,0 +1,21 @@
+'use strict';
+
+import {EventEmitter} from 'events';
+
+const CHANGE_EVENT = 'change';
+
+const BaseStore = Object.assign({}, EventEmitter.prototype, {
+  emitChange: function() {
+    this.emit(CHANGE_EVENT);
+  },
+
+  addChangeListener: function(callback) {
+    this.on(CHANGE_EVENT, callback);
+  },
+
+  removeChangeListener: function(callback) {
+    this.removeListener(CHANGE_EVENT, callback);
+  }
+});
+
+export default BaseStore;
