+++ app/stores/SearchStore.js
@@ -2,27 +2,13 @@
 
 import AppDispatcher from '../core/AppDispatcher';
 import ActionTypes from '../constants/ActionTypes';
-import {EventEmitter} from 'events';
+import BaseStore from './BaseStore';
 
-const CHANGE_EVENT = 'change';
 let _fetchingData = false;
 let _searchQuery = '';
 let _searchResults = [];
 
-
-const SearchStore = Object.assign({}, EventEmitter.prototype, {
-  emitChange: function() {
-    this.emit(CHANGE_EVENT);
-  },
-
-  addChangeListener: function(callback) {
-    this.on(CHANGE_EVENT, callback);
-  },
-
-  removeChangeListener: function(callback) {
-    this.removeListener(CHANGE_EVENT, callback);
-  },
-
+const SearchStore = Object.assign({}, BaseStore, {
   getFetchingData: function() {
     return _fetchingData;
   },
@@ -59,7 +45,6 @@ SearchStore.dispatchToken = AppDispatcher.register(function(payload) {
 });
 
 function _receiveSearchResults(schools) {
-
   _searchResults = schools.map(function(school) {
     return school.id;
   });
