+++ app/stores/UIStore.js
@@ -2,25 +2,11 @@
 
 import AppDispatcher from '../core/AppDispatcher';
 import ActionTypes from '../constants/ActionTypes';
-import {EventEmitter} from 'events';
+import BaseStore from './BaseStore';
 
-const CHANGE_EVENT = 'change';
 let _schoolTimelineYear = null;
 
-
-const UIStore = Object.assign({}, EventEmitter.prototype, {
-  emitChange: function() {
-    this.emit(CHANGE_EVENT);
-  },
-
-  addChangeListener: function(callback) {
-    this.on(CHANGE_EVENT, callback);
-  },
-
-  removeChangeListener: function(callback) {
-    this.removeListener(CHANGE_EVENT, callback);
-  },
-
+const UIStore = Object.assign({}, BaseStore, {
   getSchoolTimelineYear: function() {
     return _schoolTimelineYear;
   }
